# public/openapi.yaml
openapi: 3.0.1
info:
    title: Web3 Plugin
    description: A plugin that allows the user to generate crypto wallets and execute actions on ethereum or compatible networks. If you do not know the seedphrase, ask user for it before making requests or offer to generate a new wallet for user. If you do not know the network use Ethereum as default. Available networks are Ethereum, Polygon, Optimism, Gnosis and Linea
    version: "v1"
servers:
    - url: http://localhost:4000
paths:

    /wallets:
        post:
            operationId: generateWallet
            summary: Generate a new crypto wallet
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/walletRequest"
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/WalletInfo"
                "422":
                    description: Validation Error
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/HTTPValidationError"

    /wallets/walletaddress:
        post:
            operationId: walletAddress
            summary: Get a wallet address
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/walletAddressRequest"
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/WalletAddress"
                "422":
                    description: Validation Error
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/HTTPValidationError"

    /wallets/wallettransfer:
        post:
            operationId: walletTransfer
            summary: Send value from wallet to address
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/walletTransferRequest"
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/WalletTransfer"
                "422":
                    description: Validation Error
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/HTTPValidationError"

    /wallets/walletbalance:
        post:
            operationId: walletBalance
            summary: Get a wallet balance
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/walletBalanceRequest"
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/WalletBalance"
                "422":
                    description: Validation Error
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/HTTPValidationError"


    /contracts/deploycontract:
        post:
            operationId: deployContract
            summary: Deploy a smart contract
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/deployContractRequest"
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/deployContract"
                "422":
                    description: Validation Error
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/HTTPValidationError"

components:
    schemas:

        # wallet creation
        walletRequest:
            type: object
            properties:
                network:
                    type: string
                    description: The wallet network
        WalletInfo:
            type: object
            properties:
                seedphrase:
                    type: string
                    description: The seedphrase of wallet
                network:
                    type: string
                    description: Targeted network
                address:
                    type: string
                    description: The address of wallet

        # wallet address
        walletAddressRequest:
            type: object
            properties:
                seedphrase:
                    type: string
                    description: The wallet seedphrase
                network:
                    type: string
                    description: Targeted network
        WalletAddress:
            type: object
            properties:
                address:
                    type: string
                    description: The address of wallet

        # wallet transfer
        walletTransferRequest:
            type: object
            properties:
                seedphrase:
                    type: string
                    description: The wallet seedphrase
                network:
                    type: string
                    description: Targeted network
                address:
                    type: string
                    description: Address we are sending or transfering to
                amount:
                    type: string
                    description: Amount of value to send
        WalletTransfer:
            type: object
            properties:
                transactionId:
                    type: string
                    description: Transaction hash on-chain
                transactionLink:
                    type: string
                    description: Link to transaction on block exporer

        # wallet balances
        walletBalanceRequest:
            type: object
            properties:
                seedphrase:
                    type: string
                    description: The wallet seedphrase
                network:
                    type: string
                    description: Targeted network
        WalletBalance:
            type: object
            properties:
                balance:
                    type: string
                    description: The balance of wallet

        # deploy contract
        deployContractRequest:
            type: object
            properties:
                seedphrase:
                    type: string
                    description: The wallet seedphrase
                network:
                    type: string
                    description: Targeted network
                code:
                    type: string
                    description: Solidity code to be deployed
        deployContract:
            type: object
            properties:
                abi:
                    type: string
                    description: ABIs of a deployed smart contract
                address:
                    type: string
                    description: Address of deployed smart contract
                detailsLink:
                    type: string
                    description: Link to transaction on block exporer

        # error handling
        HTTPValidationError:
            title: HTTPValidationError
            type: object
            properties:
                errors:
                    title: errors
                    type: array
                    items:
                        $ref: "#/components/schemas/ValidationError"
        ValidationError:
            title: ValidationError
            required:
                - location
                - msg
                - path
                - type
            type: object
            properties:
                location:
                    title: Location
                    type: string
                msg:
                    title: Message
                    type: string
                path:
                    title: Path
                    type: string
                type:
                    title: Error Type
                    type: string
                